add example

+++ aprender_con_ejemplos/01.holamundo/04-holamundo.rb
@@ -2,6 +2,6 @@
 
 print("Name: "); name = gets.chomp
 print("Age: "); age = gets.to_i
-print("Height: "); height = gets.to_i
+print("Height: "); height = gets.to_f
 
 puts "El personaje #{name}, tiene #{age} años de edad y mide #{height} metros."
